CASMINST-7035 remove docs for manual CSM upgrade in 1.6 (#5510)
* CASMINST-7035 remove docs for manual CSM upgrade in 1.6

* CASMINST-7037 rename upgrade_csm_iuf_additional_products_with_iuf.md to upgrade_csm_and_additional_products_with_iuf.md

* CASMINST-7035 remove upgrade_csm_and_additonal_products.md

* CASMINST-7035: Improve "SAT and IUF" section

Improve the "SAT and IUF" section of the documentation that describes
the relationship between SAT and IUF, specifically how `sat
bootprep` is executed in IUF and how variable substitution works in
that context.

Include a procedure to obtain the `session_vars.yaml` file and
bootprep input files from the IUF activity and use them to create CFS
configurations, images, and session templates manually. This can be a
useful debugging procedure.

Include another procedure that describes how to assign the CFS
configurations and IMS images to management NCNs as a non-standard
debugging procedure. This procedure used to be used when performing an
upgrade of only CSM on a system that also contains other products that
were previously installed by IUF.

Move a section about limiting `sat bootprep run` into separate stages
into the `SAT_Bootprep.md` file because it's really more appropriate
with the general `sat bootprep` usage information.

* Apply wording improvement suggestions from code review

Co-authored-by: Ryan Haasken <77809410+haasken-hpe@users.noreply.github.com>
Signed-off-by: Lindsay Eliasen <87664908+leliasen-hpe@users.noreply.github.com>

* CASMINST-7035 address review comments and remove CSM upgrade without IUF doc

* CASMINST-7035 remove note in management_rollout

---------

Signed-off-by: Lindsay Eliasen <87664908+leliasen-hpe@users.noreply.github.com>
Co-authored-by: Ryan Haasken <ryan.haasken@hpe.com>
Co-authored-by: Ryan Haasken <77809410+haasken-hpe@users.noreply.github.com>